+++ apps/samples/buttons.cc
@@ -8,7 +8,7 @@ int main(int argc,const char*argv[]){
     cdroid::Context*ctx=&app;
     Window*w=new Window(0,0,-1,-1);
     w->setId(1);
-    w->setBackgroundColor(0xFFFF3344);
+    w->setBackgroundColor(0xFF223344);
     Drawable*d=nullptr;
     StateListDrawable*sld;
     CompoundButton*chk;
@@ -19,7 +19,7 @@ int main(int argc,const char*argv[]){
     //d=ctx->getDrawable("cdroid:drawable/btn_default.xml");
     d=ctx->getDrawable("cdroid:mipmap/textfield_default_mtrl_alpha");
     sld=dynamic_cast<StateListDrawable*>(d);
-    w->setBackgroundColor(0xFF101112);
+    btn->setBackgroundColor(0xFF332211);
     btn->setOnTouchListener([&argc](View&v,MotionEvent&e){
         const bool down=e.getAction()==MotionEvent::ACTION_DOWN;
         AnimatorSet*aset= new AnimatorSet();
@@ -32,37 +32,26 @@ int main(int argc,const char*argv[]){
         aset->start();
         return false;
     });
-    SpringAnimation spa(btn,(FloatProperty*)&SpringAnimation::SCALE_X,0.3);
+    SpringAnimation spa(btn,(FloatProperty*)&SpringAnimation::SCALE_X,0.2);
     spa.getSpring()
-        ->setStiffness(400.0f)
-        .setDampingRatio(0.5f);
+        ->setStiffness(20.0f)
+        .setDampingRatio(0.1f);
     DynamicAnimation::OnAnimationUpdateListener upls([&](DynamicAnimation& animation, float value, float velocity) {
         printf("[UPDATE] frame value=%.2f  velocity=%.2f\n", value, velocity);
     }); 
-    spa.addUpdateListener(upls);
+    DynamicAnimation::OnAnimationEndListener endls([&](DynamicAnimation& animation,bool canceled, float value, float velocity) {
+        printf("[END] frame value=%.2f  velocity=%.2f\n", value, velocity);
+    });
+    spa.addUpdateListener(upls).addEndListener(endls).start();
 
-    /*spa.addListener(std::make_shared<AnimatorListenerAdapter>(
-        [&](Animator* animation) {                                          // onAnimationStart
-            printf("[START] 弹簧动画开始\n");
-        },
-        nullptr,                                                            // onAnimationRepeat (无)
-        [&](Animator* animation) {                                         // onAnimationEnd
-            printf("[END]   弹簧动画正常结束\n");
-        },
-        [&](Animator* animation) {                                         // onAnimationCancel
-            printf("[CANCEL]弹簧动画被手动取消\n");
-        }
-    ));*/
-    spa.start();
     LOGD_IF(sld,"%p statecount=%d",sld,sld->getStateCount());
-    btn->setBackground(d);
     btn->setBackgroundTintList(ctx->getColorStateList("cdroid:color/textview"));
     btn->setTextAlignment(View::TEXT_ALIGNMENT_CENTER);
     btn->setOnClickListener([](View&v){LOGD(" Button Clicked ");});
     btn->setOnLongClickListener([](View&v)->bool{LOGD(" Button LongClicked ");return true;});
     w->addView(btn);
     btn->setId(100);
-    btn->layout(50,60,120,60);
+    btn->layout(10,60,200,60);
 
     ShapeDrawable*sd=new ShapeDrawable();
     sd->setShape(new ArcShape(0,360));