+++ src/gui/animation/dynamicanimation.cc
@@ -1,3 +1,20 @@
+/*********************************************************************************
+ * Copyright (C) [2019] [houzh@msn.com]
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+*/
 #include <animation/dynamicanimation.h>
 
 namespace cdroid{
@@ -14,7 +31,7 @@ public:                                                     \
         return ((View*)obj)->get##METHOD();                 \
     }                                                       \
 };                                                          \
-prop_##PROJ INST_##PROJ;                                    \
+static prop_##PROJ INST_##PROJ;                             \
 }                                                           \
 const FloatProperty& DynamicAnimation::PROJ =INST_##PROJ;
 
@@ -47,16 +64,15 @@ public:
         return ((FloatValueHolder*)obj)->getValue();
     }
 };
-namespace {
-    FloatValueHolderProperty FLOATVALUEHOLDER;
-}
+static  FloatValueHolderProperty FLOATVALUEHOLDER;
+
 DynamicAnimation::DynamicAnimation(FloatValueHolder* floatValueHolder) {
     mTarget = nullptr;
     mProperty = &FLOATVALUEHOLDER;
     mMinVisibleChange = MIN_VISIBLE_CHANGE_PIXELS;
 }
 
-DynamicAnimation::DynamicAnimation(void* object, FloatProperty* property) {
+DynamicAnimation::DynamicAnimation(void* object,const FloatProperty* property) {
     mTarget = object;
     mProperty = property;
     if ((mProperty == &ROTATION) || (mProperty == &ROTATION_X) || (mProperty == &ROTATION_Y)) {