+++ src/gui/animation/springanimation.cc
@@ -1,3 +1,20 @@
+/*********************************************************************************
+ * Copyright (C) [2019] [houzh@msn.com]
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+*/
 #include <animation/springanimation.h>
 namespace cdroid{
 
@@ -5,11 +22,11 @@ SpringAnimation::SpringAnimation(FloatValueHolder* floatValueHolder)
     :DynamicAnimation(floatValueHolder){
 }
 
-SpringAnimation::SpringAnimation(void*object, FloatProperty* property)
+SpringAnimation::SpringAnimation(void*object,const FloatProperty* property)
     :DynamicAnimation(object, property){
 }
 
-SpringAnimation::SpringAnimation(void* object, FloatProperty*property,float finalPosition)
+SpringAnimation::SpringAnimation(void* object,const FloatProperty*property,float finalPosition)
     :DynamicAnimation(object, property){
     mSpring = new SpringForce(finalPosition);
 }