+++ src/gui/widgetEx/wear/circularprogresslayoutcontroller.cc
@@ -1,34 +1,44 @@
+/*********************************************************************************
+ * Copyright (C) [2019] [houzh@msn.com]
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+*/
 #include <widgetEx/wear/circularprogressdrawable.h>
 #include <widgetEx/wear/circularprogresslayoutcontroller.h>
 namespace cdroid{
 
 CircularProgressLayoutController::CircularProgressLayoutController(CircularProgressLayout* layout) {
     mLayout = layout;
+    mIsTimerRunning = false;
+    mIsIndeterminate=false;
     mTotalTime = 0;
     mInterval = 0;
 }
 
-/**
- * Returns the registered {@link CircularProgressLayout.OnTimerFinishedListener}.
- */
 CircularProgressLayout::OnTimerFinishedListener CircularProgressLayoutController::getOnTimerFinishedListener() const{
     return mOnTimerFinishedListener;
 }
 
-/**
- * Sets the {@link CircularProgressLayout.OnTimerFinishedListener} to be notified when timer
- * countdown is finished.
- */
 void CircularProgressLayoutController::setOnTimerFinishedListener(const CircularProgressLayout::OnTimerFinishedListener& listener) {
     mOnTimerFinishedListener = listener;
 }
 
-/** Returns true if the progress is shown as an indeterminate spinner. */
 bool CircularProgressLayoutController::isIndeterminate() const{
     return mIsIndeterminate;
 }
 
-/** Returns true if timer is running. */
 bool CircularProgressLayoutController::isTimerRunning() const{
     return mIsTimerRunning;
 }