+++ src/gui/widgetEx/wear/curvinglayoutcallback.h
@@ -1,3 +1,20 @@
+/*********************************************************************************
+ * Copyright (C) [2019] [houzh@msn.com]
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+*/
 #ifndef __CURVING_LAYOUT_CALLBACK_H__
 #define __CURVING_LAYOUT_CALLBACK_H__
 #include <widgetEx/recyclerview/recyclerview.h>
@@ -8,7 +25,7 @@ class CurvingLayoutCallback{
 private:
     static constexpr float EPSILON = 0.001f;
 
-    Path* mCurvePath;
+    Cairo::RefPtr<cdroid::Path> mCurvePath;
     PathMeasure* mPathMeasure;
     int mCurvePathHeight;
     int mXCurveOffset;
@@ -16,8 +33,8 @@ private:
     float mCurveBottom;
     float mCurveTop;
     float mLineGradient;
-    float mPathPoints[2];
-    float mPathTangent[2];
+    double mPathPoints[2];
+    double mPathTangent[2];
     float mAnchorOffsetXY[2];
 
     RecyclerView* mParentView;